lead_capture_app/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
pytest
httpx
python-multipart
sqlalchemy
databases
aiosqlite 
requests